# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/invites/preview": {
      "post": {
        "operationId": "InvitesController_previewByBody",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePreviewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/invites/join": {
      "post": {
        "operationId": "InvitesController_joinByBody",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinByTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestCreatedDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/spaces/{id}/invites": {
      "post": {
        "operationId": "InvitesController_create",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Needs members EDIT; 403 otherwise.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateInviteResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      },
      "get": {
        "operationId": "InvitesController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/spaces/{id}/invites/{iid}/revoke": {
      "post": {
        "operationId": "InvitesController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Waiting requests on this invite become REJECTED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteSummaryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/invites/{token}/preview": {
      "get": {
        "operationId": "InvitesController_preview",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "404 INVITE_INVALID for unknown, revoked, expired or used-up invites.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePreviewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/invites/{token}/join-requests": {
      "post": {
        "operationId": "InvitesController_join",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJoinRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Idempotent per invite and Actor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestCreatedDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/join-requests/{rid}": {
      "get": {
        "operationId": "InvitesController_status",
        "parameters": [
          {
            "name": "rid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Own requests only; 404 otherwise.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/join-requests/{rid}/guardian-confirm": {
      "post": {
        "operationId": "InvitesController_guardianConfirm",
        "parameters": [
          {
            "name": "rid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Caller must be a Family guardian of the requester.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestDecisionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/me/guardian-requests": {
      "get": {
        "operationId": "InvitesController_guardianInbox",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardianRequestListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/spaces/{id}/join-requests": {
      "get": {
        "operationId": "InvitesController_listRequests",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/spaces/{id}/join-requests/{rid}/approve": {
      "post": {
        "operationId": "InvitesController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveJoinRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "409 INVITE_INVALID when the invite expired, was revoked or is used up.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestDecisionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
    "/api/v1/spaces/{id}/join-requests/{rid}/reject": {
      "post": {
        "operationId": "InvitesController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinRequestDecisionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "invites"
        ]
      }
    },
...
    "/api/v1/spaces/{id}/media/{kind}/{recordId}/{assetId}/info": {
      "get": {
        "operationId": "MediaController_info",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaInfoDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "storage"
        ]
      }
    },
    "/api/v1/spaces/{id}/media/{kind}/{recordId}/{assetId}": {
      "post": {
        "operationId": "MediaController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaStoredDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "storage"
        ]
      },
      "get": {
        "operationId": "MediaController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized, decrypted bytes; private no-store."
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "storage"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lịch Gia Đình API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fc_sid"
      }
    },
    "schemas": {
...
      "InviteTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "InvitePreviewDto": {
        "type": "object",
        "properties": {
          "space_name": {
            "type": "string"
          },
          "space_kind": {
            "type": "string",
            "enum": [
              "FAMILY",
              "GROUP"
            ]
          },
          "inviter_display_name": {
            "type": "string"
          },
          "expires_at": {
            "type": "string"
          }
        },
        "required": [
          "space_name",
          "space_kind",
          "inviter_display_name",
          "expires_at"
        ]
      },
      "JoinByTokenDto": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 50
          },
          "proposed_profile": {
            "type": "string",
            "enum": [
              "PARENT",
              "SENIOR",
              "CHILD"
            ]
          },
          "member_id": {
            "type": "string",
            "format": "uuid",
            "description": "Existing Member the requester says they are."
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "token"
        ]
      },
      "JoinRequestCreatedDto": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PENDING_GUARDIAN",
              "APPROVED",
              "REJECTED",
              "EXPIRED"
            ]
          }
        },
        "required": [
          "request_id",
          "status"
        ]
      },
      "CreateInviteDto": {
        "type": "object",
        "properties": {
          "max_uses": {
            "type": "number",
            "minimum": 1,
            "maximum": 20
          },
          "expires_in_hours": {
            "type": "number",
            "minimum": 1,
            "maximum": 168
          },
          "approval_policy": {
            "type": "string",
            "enum": [
              "APPROVAL_REQUIRED"
            ]
          },
          "proposed_role": {
            "type": "string",
            "description": "Role key suggested to the approver (not OWNER/ORGANIZER)."
          },
          "email": {
            "type": "string",
            "description": "Addressee; the invite then shows in their GET /me/invitations."
          }
        },
        "required": [
          "max_uses",
          "expires_in_hours",
          "approval_policy"
        ]
      },
      "CreateInviteResponseDto": {
        "type": "object",
        "properties": {
          "email_delivery": {
            "type": "string",
            "enum": [
              "SENT",
              "NOT_CONFIGURED",
              "FAILED"
            ]
          },
          "invite_id": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string",
            "description": "Shown once; only its hash is stored."
          },
          "url": {
            "type": "string",
            "example": "/tham-gia/?token=…"
          },
          "expires_at": {
            "type": "string"
          }
        },
        "required": [
          "invite_id",
          "token",
          "url",
          "expires_at"
        ]
      },
      "InviteSummaryDto": {
        "type": "object",
        "properties": {
          "invite_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "EXPIRED",
              "USED_UP",
              "REVOKED"
            ]
          },
          "max_uses": {
            "type": "number"
          },
          "uses": {
            "type": "number"
          },
          "expires_at": {
            "type": "string"
          },
          "proposed_role": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "invite_id",
          "status",
          "max_uses",
          "uses",
          "expires_at",
          "proposed_role",
          "email",
          "created_at"
        ]
      },
      "InviteListDto": {
        "type": "object",
        "properties": {
          "invites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InviteSummaryDto"
            }
          }
        },
        "required": [
          "invites"
        ]
      },
      "CreateJoinRequestDto": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 50
          },
          "proposed_profile": {
            "type": "string",
            "enum": [
              "PARENT",
              "SENIOR",
              "CHILD"
            ]
          },
          "member_id": {
            "type": "string",
            "format": "uuid",
            "description": "Existing Member the requester says they are."
          }
        },
        "required": [
          "display_name"
        ]
      },
      "JoinRequestStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PENDING_GUARDIAN",
              "APPROVED",
              "REJECTED",
              "EXPIRED"
            ]
          },
          "space_id": {
            "type": "string",
            "format": "uuid",
            "description": "Only once APPROVED."
          }
        },
        "required": [
          "status"
        ]
      },
      "JoinRequestDecisionDto": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "member_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "request_id",
          "status"
        ]
      },
      "GuardianRequestDto": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "display_name": {
            "type": "string"
          },
          "space_name": {
            "type": "string"
          },
          "space_kind": {
            "type": "string",
            "enum": [
              "FAMILY",
              "GROUP"
            ]
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "display_name",
          "space_name",
          "space_kind",
          "created_at"
        ]
      },
      "GuardianRequestListDto": {
        "type": "object",
        "properties": {
          "requests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GuardianRequestDto"
            }
          }
        },
        "required": [
          "requests"
        ]
      },
      "JoinRequestSummaryDto": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "invite_id": {
            "type": "string",
            "format": "uuid"
          },
          "display_name": {
            "type": "string"
          },
          "proposed_profile": {
            "type": "string",
            "nullable": true
          },
          "member_id": {
            "type": "string",
            "nullable": true
          },
          "proposed_role": {
            "type": "string",
            "nullable": true
          },
          "device_label": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PENDING_GUARDIAN",
              "APPROVED",
              "REJECTED",
              "EXPIRED"
            ]
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "invite_id",
          "display_name",
          "proposed_profile",
          "member_id",
          "proposed_role",
          "device_label",
          "status",
          "created_at"
        ]
      },
      "JoinRequestListDto": {
        "type": "object",
        "properties": {
          "requests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JoinRequestSummaryDto"
            }
          }
        },
        "required": [
          "requests"
        ]
      },
      "NewMemberDto": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 50
          },
          "relationship": {
            "type": "string",
            "enum": [
              "FATHER",
              "MOTHER",
              "SON",
              "DAUGHTER",
              "GRANDFATHER",
              "GRANDMOTHER",
              "GUARDIAN",
              "OTHER"
            ]
          },
          "profile": {
            "type": "string",
            "enum": [
              "PARENT",
              "SENIOR",
              "CHILD"
            ]
          }
        },
        "required": [
          "display_name",
          "relationship",
          "profile"
        ]
      },
      "ApproveJoinRequestDto": {
        "type": "object",
        "properties": {
          "role_key": {
            "type": "string"
          },
          "member_id": {
            "type": "string",
            "format": "uuid",
            "description": "Link to this existing Member (exclusive with new_member)."
          },
          "new_member": {
            "description": "Create a Member for the requester (exclusive with member_id).",
            "allOf": [
              {
                "$ref": "#/components/schemas/NewMemberDto"
              }
            ]
          }
        },
        "required": [
          "role_key"
        ]
      },
...
      "MediaInfoDto": {
        "type": "object",
        "properties": {
          "size": {
            "type": "number"
          },
          "mime": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          }
        },
        "required": [
          "size",
          "mime",
          "sha256"
        ]
      },
      "MediaStoredDto": {
        "type": "object",
        "properties": {
          "stored": {
            "type": "boolean"
          }
        },
        "required": [
          "stored"
        ]
      },
...
    }
  }
}